refactor: interview page 관련 components 폴더 구조 정리
