+++ src/app/interview/page.module.scss
@@ -19,6 +19,21 @@
   padding: 32px 80px;
 }
 
+.simulation_layout {
+  min-width: 620px;
+  padding: 40px 0px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  border-radius: 41px;
+  border: 1px solid #5351a5;
+  background: #222844;
+  width: 100%;
+  height: 100%;
+}
+
 @media (max-width: 1280px) {
   .layout {
     height: calc(100vh - 63px);
