--- src/components/interview/index.ts
@@ -1,3 +1,0 @@
-export { Ready } from '@/components/interview/Ready';
-export { Start } from '@/components/interview/Start';
-export { End } from '@/components/interview/End';
