+++ src/widgets/interview/End/End.module.scss
