+++ src/widgets/interview/End/index.tsx
