+++ src/widgets/interview/Ready/Ready.module.scss
