+++ src/widgets/interview/Ready/index.tsx
