--- src/widgets/interview/Simulation/index.module.scss
@@ -1,14 +1,0 @@
-.layout {
-  min-width: 620px;
-  padding: 40px 0px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-  border-radius: 41px;
-  border: 1px solid #5351a5;
-  background: #222844;
-  width: 100%;
-  height: 100%;
-}
