+++ src/widgets/interview/index.ts
@@ -1,2 +1,4 @@
-export { Simulation } from '@/widgets/interview/Simulation';
 export { Stacks } from '@/widgets/interview/Stacks';
+export { Ready } from '@/widgets/interview/Ready';
+export { Start } from '@/widgets/interview/Start';
+export { End } from '@/widgets/interview/End';
